+++ navigation/AppNavigator.js
@@ -5,7 +5,7 @@ import { FontAwesome } from "react-native-vector-icons";
 import { createStackNavigator } from "react-navigation-stack";
 import MainTabNavigator from "./MainTabNavigator";
 
-import SignUp from "../screens/AuthScreen";
+import SignUp from "../screens/SignUpScreen";
 import SignIn from "../screens/SignInScreen";
 
 const headerStyle = {
